backend/requirements.txt
fastapi
sqlalchemy
passlib[bcrypt]
python-multipart
python-jose[cryptography]
python-dotenv
openai
pydantic
fastapi-mail
uvicorn
psycopg2-binary
